// frontend/src/api/openapi-types.ts
export interface components {
  schemas: {
// ...
    CalendarOut: {
      /** Name */
      name: string;
      /** Color */
      color?: string | null;
      /** Description */
      description?: string | null;
      /** Id */
      id: string;
      /** Owner Id */
      owner_id?: string | null;
      /**
       * Created At
       * Format: date-time
       */
      created_at: string;
      /**
       * Updated At
       * Format: date-time
       */
      updated_at: string;
    };
// ...
    EventOut: {
      /** Title */
      title: string;
      /** Description */
      description?: string | null;
      /**
       * Start
       * Format: date-time
       */
      start: string;
      /**
       * End
       * Format: date-time
       */
      end: string;
      /**
       * All Day
       * @default false
       */
      all_day: boolean | null;
      /** Id */
      id: string;
      /** Calendar Id */
      calendar_id: string;
      /**
       * Created At
       * Format: date-time
       */
      created_at: string;
      /**
       * Updated At
       * Format: date-time
       */
      updated_at: string;
    };
    /** EventUpdate */
    EventUpdate: {
      /** Title */
      title?: string | null;
      /** Description */
      description?: string | null;
      /** Start */
      start?: string | null;
      /** End */
      end?: string | null;
      /** All Day */
      all_day?: boolean | null;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}